pangolin: Introduce ShellService
This allows us to communicate with shell without context and to use
shell specific stuff in services (more on this later)
