+++ lib/components/overlays/search/search_overlay.dart
@@ -18,8 +18,8 @@ import 'package:dahlia_shared/dahlia_shared.dart';
 import 'package:flutter/material.dart';
 import 'package:pangolin/components/overlays/search/search_tile.dart';
 import 'package:pangolin/components/overlays/search/searchbar.dart';
-import 'package:pangolin/components/shell/shell.dart';
 import 'package:pangolin/services/search.dart';
+import 'package:pangolin/services/shell.dart';
 import 'package:pangolin/utils/data/globals.dart';
 import 'package:pangolin/widgets/surface/surface_layer.dart';
 import 'package:xdg_desktop/xdg_desktop.dart';
