+++ lib/components/overlays/tray_overlay.dart
@@ -4,9 +4,9 @@ import 'package:dahlia_shared/dahlia_shared.dart';
 import 'package:dbus/dbus.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:pangolin/components/shell/shell.dart';
 import 'package:pangolin/services/dbus/image.dart';
 import 'package:pangolin/services/dbus/status_item.dart';
+import 'package:pangolin/services/shell.dart';
 import 'package:pangolin/services/tray.dart';
 import 'package:pangolin/services/wm.dart';
 import 'package:pangolin/widgets/context_menu.dart';
