+++ lib/components/taskbar/tray.dart
@@ -1,8 +1,8 @@
 import 'package:dahlia_shared/dahlia_shared.dart';
 import 'package:flutter/material.dart';
 import 'package:pangolin/components/overlays/tray_overlay.dart';
-import 'package:pangolin/components/shell/shell.dart';
 import 'package:pangolin/components/taskbar/taskbar_element.dart';
+import 'package:pangolin/services/shell.dart';
 import 'package:pangolin/services/tray.dart';
 
 class TrayMenuButton extends StatelessWidget
@@ -17,7 +17,7 @@ class TrayMenuButton extends StatelessWidget
       height: 48.0,
       child: ValueListenableBuilder<bool>(
         valueListenable:
-            Shell.of(context).getShowingNotifier(TrayMenuOverlay.overlayId),
+            ShellService.current.getShowingNotifier(TrayMenuOverlay.overlayId),
         builder: (context, showing, _) {
           return Padding(
             padding: EdgeInsets.only(
